% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniciar_geobox.R
\name{iniciar_geobox}
\alias{iniciar_geobox}
\title{Executa o Geobox}
\usage{
iniciar_geobox(...)
}
\arguments{
\item{...}{A series of options to be used inside the app.}
}
\description{
Executa o Geobox
}
